src/repositories/__init__.py
import os

from dotenv import load_dotenv

load_dotenv()

user = os.getenv("DB_USER")
password = os.getenv("DB_PASSWORD")
host = os.getenv("DB_HOST")
port = os.getenv("DB_PORT")
db_name = os.getenv("DB_NAME")
URL = f"postgresql+psycopg://{user}:{password}@{host}:{port}/{db_name}"
