requirements.txt
langchain-nomic
langchain-community
tiktoken
langchainhub
chromadb
langchain
langgraph
tavily-python
gpt4all
jupyter
psycopg2-binary
pgvector
psycopg[binary,pool]

langchain-huggingface
langchain-core
langchain-ollama
langchain-chroma
langchain-postgres
langchain-cohere
sentence-transformers

playwright
unstructured

python-dotenv
python-multipart
email_validator
fastapi
uvicorn
passlib
bcrypt
pydantic
cryptography
python-jose
sqlalchemy
